+++ backend/src/main/java/ruan/eloy/backend/dto/SignUpRequest.java
@@ -1,4 +1,4 @@
-package ruan.eloy.backend.dtos;
+package ruan.eloy.backend.dto;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
